Remove SQL utils and rename modules
