+++ bgbb/utils.py
