+++ bgbb/vectorized.py
