+++ tests/test_vectorized.py
@@ -1,7 +1,7 @@
 # import pytest
 import numpy as np
 from scipy.special import betaln, logsumexp
-from bgbb.numba_special import nb_lbeta, nb_lbeta_vec2, nb_logsumexp
+from bgbb.vectorized import nb_lbeta, nb_lbeta_vec2, nb_logsumexp
 
 
 def test_nb_lbeta(niter=1000):
